python_ingestion/pyproject.toml
[project]
name = "python_ingestion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckdb>=1.1.3",
    "pandas>=2.2.3",
    "pyarrow>=19.0.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
]
